Order detail view by customer.

+++ BookStoreWebsite/src/com/bookstore/controller/frontend/shopingcart/ShopingCart.java
@@ -35,8 +35,8 @@ public class ShopingCart {
 		return total;
 	}
 	
-	public double getTotalAmount() {
-		double total = 0.0f;
+	public float getTotalAmount() {
+		float total = 0.0f;
 		
 		Iterator<Book> iterator = cart.keySet().iterator();
